requirements.txt
boto3
requests
minio
pydantic
pydantic-settings
python-dotenv
pymupdf 
python-docx 
unstructured[all-docs] 
pandas 
numpy 
spacy 
langchain 
sentence-transformers
pdfminer.six
pi_heif
langchain_community
PyPDF2
google-genai
ollama
chromadb
InstructorEmbedding